fix commit errors but need to fix get deque methods since its confusing

+++ src/main/java/com/hobbes/wstore/DiskDataRange.java
@@ -8,7 +8,7 @@ class DiskDataRange extends DataRange {
     }
 
     @Override
-    public int getData(byte[] buf, int pos) {
+    public long getData(byte[] buf, int pos) {
 	// Do nothing for now
 	return 0;
     }
